Orientation change calculates mean / deviation sequences.
This is not yet visualized, but can be traced in StateCommand.

+++ bin-debug/assets/data/sql/create/CreateTableTasks.sql
@@ -2,9 +2,9 @@ CREATE TABLE main.tasks
 (
 	taskId INTEGER PRIMARY KEY AUTOINCREMENT,
 	name String NOT NULL,
-    mean INT,
-    deviation INT,
-    optimistic INT,
-    realistic INT,
-    pessimistic INT
+    mean FLOAT,
+    deviation FLOAT,
+    optimistic FLOAT,
+    realistic FLOAT,
+    pessimistic FLOAT
 )
